+++ app/src/main/java/com/rick/posadka/data/PosadkaRepository.kt
@@ -12,13 +12,13 @@ class PosadkaRepository @Inject constructor(
     private val posadkaShaftDao: PosadkaShaftDao
 ) {
 
-    suspend fun getPosadkaShaft(size: Int, name: String): Flow<PosadkaShaft> {
+    suspend fun getPosadkaShaft(size: Float, name: String): Flow<PosadkaShaft> {
         val queryString = "%${name.replace(' ', '%')}%"
         return posadkaShaftDao.getShaftByClass(querySize = size, queryString =queryString)
 
     }
 
-    suspend fun getPosadkaHole(size: Int, name: String): Flow<PosadkaHole> {
+    suspend fun getPosadkaHole(size: Float, name: String): Flow<PosadkaHole> {
         val queryString = "%${name.replace(' ', '%')}%"
         return posadkaHoleDao.getHoleByClass(querySize = size, queryString = name)
     }
