+++ app/src/main/java/com/rick/posadka/model/PosadkaShaft.kt
@@ -9,7 +9,7 @@ data class PosadkaShaft(
     @ColumnInfo(name = "id")
     @PrimaryKey(autoGenerate = true) val id : Int,
     @ColumnInfo(name = "size")
-    val size: Int,
+    val size: Float,
     @ColumnInfo(name = "prev_size")
     val prevSize: Int,
     @ColumnInfo(name = "class")
